pyproject.toml
[project]
name = "move-to-happy"
version = "0.1.0"
description = "A collection of scripts to help me move to happy"
authors = [
    {name = "david-mth"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = ["boto3 (>=1.42.51,<2.0.0)", "pandas (>=2.2,<3.0)", "pyathena (>=3.27.2,<4.0.0)", "awswrangler (>=3.15.1,<4.0.0)", "sagemaker (>=2.232,<3.0.0)", "numpy (>=2.0,<3.0)", "python-dotenv (>=1.2.1,<2.0.0)", "requests (>=2.31,<3.0)", "openpyxl (>=3.1.5,<4.0.0)", "anthropic (>=0.84.0,<0.85.0)", "faiss-cpu (>=1.13.2,<2.0.0)", "sentence-transformers (>=5.2.3,<6.0.0)", "tiktoken (>=0.12.0,<0.13.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "move_to_happy", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.9"
mypy = "^1.14"
pre-commit = "^4.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py313"
line-length = 88
src = ["src", "tests"]
extend-exclude = ["*.ipynb"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]

[tool.ruff.lint.per-file-ignores]
"scripts/**/*.py" = ["N999"]  # Numbered scripts use invalid module names by design

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
